Rename exponential_decay as standalone/model
